Adding confirmation prompt to creating database

+++ mysql_db/create_database.py
@@ -2,22 +2,26 @@ import configparser
 from pathlib import Path
 import mysql.connector
 
-config = configparser.ConfigParser()
-config.read(Path.cwd().joinpath("config.ini"))
+confirm = input("This will delete the entire database and recreate it. If you want to proceed, enter \"confirm\": ")
 
-mydb = mysql.connector.connect(
-   host=config["mysql"]["host"], 
-   user=config["mysql"]["user"], 
-   password=config["mysql"]["password"])
+if confirm.lower() == "confirm":
 
-mycursor = mydb.cursor()
+   config = configparser.ConfigParser()
+   config.read(Path.cwd().joinpath("config.ini"))
 
-mycursor.execute("DROP DATABASE IF EXISTS nascar;")
-mycursor.execute("CREATE DATABASE nascar;")
-mydb.database = "nascar" # type: ignore
+   mydb = mysql.connector.connect(
+      host=config["mysql"]["host"],
+      user=config["mysql"]["user"],
+      password=config["mysql"]["password"])
 
-for script in Path.cwd().joinpath("mysql_db", "sql_scripts").glob("*.sql"):
-   print(script)
-   sql = script.read_text()
-   print(sql)
-   mycursor.execute(sql)
+   mycursor = mydb.cursor()
+
+   mycursor.execute("DROP DATABASE IF EXISTS nascar;")
+   mycursor.execute("CREATE DATABASE nascar;")
+   mydb.database = "nascar" # type: ignore
+
+   for script in Path.cwd().joinpath("mysql_db", "sql_scripts").glob("*.sql"):
+      print(f"Running script in {script}")
+      sql = script.read_text()
+      print(sql)
+      mycursor.execute(sql)
